+++ client/components/grade.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+class Grade extends React.Component {
+  render() {
+    return (
+      <tr>
+        <td>{this.props.grade.name}</td>
+        <td>{this.props.grade.course}</td>
+        <td>{this.props.grade.grade}</td>
+      </tr>
+    );
+  }
+}
+
+export default Grade;
