+++ client/components/header.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Header extends React.Component {
+  render() {
+    return (
+      <header>
+        <h1>Student Grade Table</h1>
+      </header>
+    );
+  }
+}
+
+export default Header;
